+++ plugins/nexus-repository-manager/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @janus-idp/backstage-plugin-nexus-repository-manager [1.4.28](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-nexus-repository-manager@1.4.27...@janus-idp/backstage-plugin-nexus-repository-manager@1.4.28) (2024-03-04)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.7.5
+
 ## @janus-idp/backstage-plugin-nexus-repository-manager [1.4.27](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-nexus-repository-manager@1.4.26...@janus-idp/backstage-plugin-nexus-repository-manager@1.4.27) (2024-02-27)