+++ plugins/web-terminal/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @janus-idp/backstage-plugin-web-terminal [1.3.7](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-web-terminal@1.3.6...@janus-idp/backstage-plugin-web-terminal@1.3.7) (2024-03-04)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.7.5
+
 ## @janus-idp/backstage-plugin-web-terminal [1.3.6](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-web-terminal@1.3.5...@janus-idp/backstage-plugin-web-terminal@1.3.6) (2024-02-27)